Workaround-fix for KT-4197 Missed warning for function literal clashed with parameter
IDEA does not display warnings if there is an error in the same position. Thus the DANGLING_FUNCTION_LITERAL_ARGUMENT_SUSPECTED was reported but not visible in Editor,
if the same lambda was marked with e.g. TOO_MANY_ARGUMENTS.

As a workaround solution I've added hint text to error messages for errors that may occur in this position

 #KT-4197 Fixed